update bootstrap.py, removed --install option

+++ bootstrap.py
@@ -35,13 +35,11 @@ import sys
 
 @click.command()
 @click.option('--bootstrap/--no-bootstrap', default=True, help="Don't bootstrap Python dependencies")
-@click.option('--install', is_flag=True, help="Install PPYM after bootstrapping "
-    "and clean up the bootstrapped nodepy_modules/ directory.")
 @click.option('-g', '--global', 'global_', is_flag=True, help="Install PPYM globally.")
 @click.option('--root', is_flag=True, help="Install PPYM in the Python root folder.")
 @click.option('-U', '--upgrade', is_flag=True, help="Uninstall previous versions instead of skipping the new version.")
 @click.option('--develop', is_flag=True, help="If --install, install in development mode.")
-def main(bootstrap, install, global_, root, upgrade, develop):
+def main(bootstrap, global_, root, upgrade, develop):
   """
   Bootstrap the PPYM installation.
   """
@@ -62,27 +60,27 @@ def main(bootstrap, install, global_, root, upgrade, develop):
       print('error: Pip installation failed')
       sys.exit(res)
 
-  if install:
-    # This is necessary on Python 2.7 (and possibly other versions) as
-    # otherwise importing the newly installed Python modules will fail.
-    sys.path_importer_cache.clear()
+  # This is necessary on Python 2.7 (and possibly other versions) as
+  # otherwise importing the newly installed Python modules will fail.
+  sys.path_importer_cache.clear()
 
-    print("Installing PPYM {} ...".format('globally' if global_ else 'locally'))
-    cmd = ['install']
-    if upgrade:
-      cmd.append('--upgrade')
-    if global_:
-      cmd.append('--global')
-    if root:
-      cmd.append('--root')
-    if develop:
-      cmd.append('--develop')
-    cmd.append(__directory__)
-    require('./index').main(cmd, standalone_mode=False)
+  print("Installing PPYM {} ...".format('globally' if global_ else 'locally'))
+  cmd = ['install']
+  if upgrade:
+    cmd.append('--upgrade')
+  if global_:
+    cmd.append('--global')
+  if root:
+    cmd.append('--root')
+  if develop:
+    cmd.append('--develop')
+  cmd.append(__directory__)
+  require('./index').main(cmd, standalone_mode=False)
 
-    if not existed_before and bootstrap and os.path.isdir('nodepy_modules'):
-      print('Cleaning up bootstrap modules directory ...')
-      shutil.rmtree('nodepy_modules')
+  local = (not global_ and not root)
+  if not local and not existed_before and bootstrap and os.path.isdir('nodepy_modules'):
+    print('Cleaning up bootstrap modules directory ...')
+    shutil.rmtree('nodepy_modules')
 
 
 main()
